#572
Hello @apu52 sir!!
we @Shivani756 and @litesh1123 have completed the feature of the  introduction of google maps. #572
But the api key verification is taking time and it says that we'll need to pay if the usage exceeds a particular limit . I'm sharing the screenshots with you and request you to generate the key from your side and replace the key in the maps.html file.
rest of the code is completed.....
after that if there will be issue then we will be assisting you

Co-Authored-By: litesh1123 <139573484+litesh1123@users.noreply.github.com>